# Conflicts:
#	src/main/java/io/swagger/api/AccountsApiController.java
#	src/main/java/io/swagger/api/LoginApiController.java
#	src/main/java/io/swagger/dao/AccountRepository.java
#	src/main/java/io/swagger/model/LoginDetails.java
#	src/main/java/io/swagger/model/UserCredentials.java
#	src/main/java/io/swagger/service/AccountService.java

+++ src/main/java/io/swagger/service/AccountService.java
@@ -4,7 +4,6 @@ import io.swagger.dao.AccountRepository;
 import io.swagger.filter.Filter;
 import io.swagger.model.Account;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
@@ -33,32 +32,39 @@ public class AccountService {
 
 
     public Account getSpecificAccount(String iBan) {
-        Account specificAccount;
-        try {
-            specificAccount = accountRepository.findById(iBan).get();
-            return specificAccount;
-        } catch (Exception e) {
-            logger.warning("Failed to get accounts " + e.getMessage());
+        if (iBan.equals(""))
+        {
+            logger.warning("Invalid IBAN provided.");
+            throw new IllegalArgumentException("Invalid IBAN provided.");
         }
-        return new Account();
+
+        Account account = accountRepository.findById(iBan).get();
+        if (account == null)
+        {
+            logger.warning("Account "+ iBan + " does not exist.");
+            throw new IllegalArgumentException("Account "+ iBan + " does not exist.");
+        }
+        return account;
     }
 
     public void deleteAccount(String iBan) {
-        try {
+        if (accountRepository.existsById(iBan))
+        {
             accountRepository.deleteById(iBan);
-        } catch (Exception e) {
-            logger.warning("Failed to delete account " + e.getMessage());
+        } else {
+            logger.warning("Failed to delete account:  " + iBan + " does not exist.");
+            throw new IllegalArgumentException("Failed to delete account: " + iBan + " does not exist.");
         }
     }
 
     public Account editAccount(String iBan, Account updatedAccount) {
-        try {
-            accountRepository.save(updatedAccount);
-            return accountRepository.findById(iBan).get();
-        } catch (Exception e) {
-            logger.warning("Failed to edit account " + e.getMessage());
+        if (iBan.equals("") || updatedAccount == null)
+        {
+            logger.warning("Invalid IBAN or account provided.");
+            throw new IllegalArgumentException("Invalid IBAN or account provided.");
         }
-        return new Account();
+        accountRepository.save(updatedAccount);
+        return accountRepository.findById(iBan).get();
     }
 
     private Iterable<Account> fillResponse(Filter filter) {
